fix: Override number of stop bits in case of error
On Windows, the serial driver reports a setting of 1.5 stop bits in crate serialport (https://crates.io/crates/serialport). This is not supported by the crate so an error is returned. For unknown reasons the stop bit setting can not be overwritten before establishing a connectino to the serial port. Trying to override it using the set_stop_bits() function the changes are just not applied. So, this workaround sets the stop bits to 1 in case an error is returned.